Add files via upload
a web server with the ESP32 that displays a
web page to control a stepper motor. The web page allows you to insert the
number of steps and select clockwise or counterclockwise direction. Additionally, it
also shows whether the motor is currently spinning or if it is stopped. The
communication between the client and the server is achieved via WebSocket
protocol. All clients are updated with the current motor state.

https://randomnerdtutorials.com/stepper-motor-esp32-websocket/